packages/neckenml-analyzer/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neckenml-analyzer"
version = "0.3.0"
description = "Audio analysis for Swedish folk music using Essentia (AGPL licensed)"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "AGPL-3.0-or-later"}
authors = [
    {name = "NeckenML Contributors"}
]
keywords = ["audio", "music", "folk", "swedish", "analysis", "essentia"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "neckenml-core>=0.3.0",
    "essentia-tensorflow>=2.1b6.dev0",
    "librosa>=0.9.0",
    "madmom>=0.16.0",
    "soundfile>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/svnoak/neckenml-analyzer"
Repository = "https://github.com/svnoak/neckenml-analyzer"
Issues = "https://github.com/svnoak/neckenml-analyzer/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39']
